refactor: User 관련해 int 사용된 부분 long으로 바꿈

+++ src/main/java/kr/codesqaud/cafe/domain/User.java
@@ -1,8 +1,8 @@
 package kr.codesqaud.cafe.domain;
 
 public class User {
-    private static int userNumFactory = 0;
-    private final int userNum;
+    private static long userNumFactory = 0;
+    private final long userNum;
     private String userId;
     private String password;
     private String email;       // form에서 사용된 name에 맞춤
